refactored to match updated entities

+++ src/main/java/com/example/dependencyinjectiontask/CourseService.java
@@ -1,6 +1,6 @@
 package com.example.dependencyinjectiontask;
 
-import jakarta.persistence.EntityNotFoundException;
+import javax.persistence.*;
 import org.example.CourseDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -48,7 +48,7 @@ public class CourseService {
     }
 
     public void updateCourse(Course course) {
-        if (courseRepository.existsById(course.getId())) {
+        if (courseRepository.existsById((int) course.getId())) {
             courseRepository.save(course);
         }
     }
